+++ packages/plugin-packages/mashroom-storage-provider-mongodb/test/MashroomStorageCollectionMongoDB.test.ts
@@ -69,8 +69,8 @@ describe('MashroomStorageCollectionMongoDB', () => {
 
         expect(result1).toBeTruthy();
         expect(result2).toBeTruthy();
-        expect(result1.length).toBe(3);
-        expect(result2.length).toBe(1);
+        expect(result1.result.length).toBe(3);
+        expect(result2.result.length).toBe(1);
     });
 
     it('supports filter operators', async () => {
@@ -92,13 +92,17 @@ describe('MashroomStorageCollectionMongoDB', () => {
         expect(result2).toBeTruthy();
         expect(result3).toBeTruthy();
         expect(result4).toBeTruthy();
-        expect(result1.length).toBe(1);
-        expect(result1[0].b).toBe(2);
-        expect(result2.length).toBe(3);
-        expect(result2[2].b).toBe(2);
-        expect(result3.length).toBe(3);
-        expect(result3[0].b).toBe(2);
-        expect(result4.length).toBe(1);
+        expect(result1.result.length).toBe(1);
+        expect(result1.totalCount).toBe(1);
+        expect(result1.result[0].b).toBe(2);
+        expect(result2.result.length).toBe(3);
+        expect(result2.totalCount).toBe(3);
+        expect(result2.result[2].b).toBe(2);
+        expect(result3.result.length).toBe(3);
+        expect(result3.totalCount).toBe(3);
+        expect(result3.result[0].b).toBe(2);
+        expect(result4.result.length).toBe(1);
+        expect(result4.totalCount).toBe(1);
     });
 
     it('updates all properties of an existing property with updateOne', async () => {
@@ -145,8 +149,8 @@ describe('MashroomStorageCollectionMongoDB', () => {
         const result2 = await storage.find({b: 1});
 
         expect(deleteResult.deletedCount).toBe(1);
-        expect(result1.length).toBe(1);
-        expect(result2.length).toBe(0);
+        expect(result1.result.length).toBe(1);
+        expect(result2.result.length).toBe(0);
     });
 
     it('deletes all existing items with deleteMany', async () => {
@@ -162,9 +166,10 @@ describe('MashroomStorageCollectionMongoDB', () => {
         const result2 = await storage.find({a: 1});
 
         expect(deleteResult.deletedCount).toBe(3);
-        expect(result1.length).toBe(3);
-        expect(result2.length).toBe(0);
+        expect(result1.result.length).toBe(3);
+        expect(result2.result.length).toBe(0);
     });
+
 
 });