+++ packages/plugin-packages/mashroom-storage/type-definitions/index.js
@@ -14,6 +14,11 @@ export type MashroomStorageSort<T: MashroomStorageRecord> = {
     [$Keys<MashroomStorageObject<T>>]: 'asc' | 'desc',
 }
 
+export type MashroomStorageSearchResult<T> = {
+    result: Array<MashroomStorageObject<T>>;
+    totalCount: number;
+}
+
 export type MashroomStorageUpdateResult = {
     modifiedCount: number,
 };
@@ -29,7 +34,7 @@ export interface MashroomStorageCollection<T: MashroomStorageRecord> {
     /**
      * Find all items that match given filter. The filter supports a subset of Mongo's filter operations (like $gt, $regex, ...).
      */
-    find(filter?: MashroomStorageObjectFilter<T>, limit?: number, skip?: number, sort?: MashroomStorageSort<T>): Promise<Array<MashroomStorageObject<T>>>;
+    find(filter?: MashroomStorageObjectFilter<T>, limit?: number, skip?: number, sort?: MashroomStorageSort<T>): Promise<MashroomStorageSearchResult<T>>;
     /**
      * Return the first item that matches the given filter or null otherwise.
      */